restructured directory

+++ app/src/main/java/com/example/productlister/domain/model/Product.kt
@@ -2,7 +2,6 @@ package com.example.productlister.domain.model
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
-import java.time.LocalDateTime
 import java.util.Date
 
 @Entity(tableName = "consumer_product")
@@ -10,7 +9,7 @@ data class Product(
     @PrimaryKey(autoGenerate = true) val id: Int?,
     val name: String,
     val price: Int,
-    val pricePerG: Float,
+    val pricePerG: Double,
     val netWt: Int,
     val img1: String?,
     val img2: String?,
